mzconduct: Put error logic and handling in materialize.errors
This creates two trees of errors: MzConfigurationError and MzRuntimeError, the first for
things that are programmer error, and the second for runtime errors that cannot be
statically detected.
